gollvm: driver code reorganization
Split the existing driver code into two subdirectories, one with a
driver utilities library and one with the driver executable itself
(which links against the utilities library). This is intended to make
it possible write unit tests for code in the utilities library. Add a
stub driver utilities unit test (contents to be added later).

Change-Id: I9e3841b85bae10c918552cb3ef70297243ae970d