gollvm: roll out new driver functionality for assembly phase
Transition over to using the new helper classes (Driver, Compilation,
etc) for the assembly portion of the compilation. The main compile (Go
to assembly) is still being done as before. There are some temporary
hacks needed in the way that output Artifacts are created; these will
go away in a subsequent patch.

Change-Id: I9aae0e7ee1846286029f3181111ed451fe369aeb