gollvm: preliminary support for link step of compilation
Add basic support for linking Go programs. This involves locating the
correct set of files/paths and passing the correct arguments to the
linker. Note: this patch brings us to a point where we can handle the
linking steps as part of the libgo/gotools builds, but we're not quite
there yet for complete "go build" use. In particular, this patch
includes a short-term strategy for locating the libgo directory
relative to the installed path of llvm-goc binary. Better install
support will arrive in a forthcoming patch.

Change-Id: Ief1f3d8d37c24e28fd421c0b77a9e23af2f816ac