gollvm: resolve arg[0] path in -v/-### trace output
The go command takes advantage of the -### compiler output to
derive the path to the compiler executable, as part of computing
a checksum / buildid for the compilation. To help with this,
make sure that the -v / -### output emits a resolved executable
path and not just the argv[0] value. This patch also includes
some additional -### bugfixes.

Change-Id: Iab5faa7da57c479e1fb083d4d3797e5ff3a3c93f