gollvm: add cmake controls for go compiler dependence
In typical use cases, we want to cmake to force a rebuild of all
objects built from Go source if the compiler binary changes For
compiler developers, however, this dependence can be a hassle-- when
iterating on compiler changes, it is annoying to have each tiny change
force a library rebuild.

To help with this, create a new (unsafe) DISABLE_LIBGO_GOC_DEP cmake
control variable that can be enabled by compiler developers to break
the dependence on the compiler for all llvm-goc compiled objects. If
DISABLE_LIBGO_GOC_DEP, objects depend on a token file that gets
created once the first copy of the compiler is built (the token file
is not updated when the compiler is rebuilt).

Change-Id: I5267c3106abd79db822e0ab2b1ac1387833969a5
Reviewed-on: https://go-review.googlesource.com/112355
Reviewed-by: Cherry Zhang <cherryyz@google.com>