driver, bridge: add stack map support machinary
Add -enable-gc={0,1} command line flag for enabling GC (stack
map) support. Default to false for now.

When the GC support is enabled, the driver will incorporate GC
related backend passes. And the bridge will attach the GC
strategy to each function, as well as attach a dummy personality
function to functions that don't have one, to ensure the
exception table is always generated.

Change-Id: I86dfd4a2c79bf00305e4301103a240b053060f17