gollvm: accept + ignore some additional command line flags
Change the command line parser to accept (and ignore) a collection of
flags supported by gccgo that are used as part of the gccgo torture
test and regression test. These include fno-diagnostics-show-caret,
-fno-show-column, -fbounds-check, -pedantic-errors, and others. Also
included on the list are -flto and -fuse_linker_plugin (it is worth
noting that at some point we want to support some form of LTO).

Change-Id: Ia084aadf662b7d2016a038f5be010fe32ff7eed5