gollvm: fix issues with generation of package search path
Rework the code that computes the Go package search path to insure
that we're being compatible with the scheme used by gccgo. After
looking in -I dirs, for -L/-B args we want to first look in
go/<version> and go/<version>/<triple> subdirs, and finally in the top
level dir itself.

Change-Id: I9af37f504698211756876416e6baa278bb34eec4