gollvm: fix problem with generation of package search path
At some point as part of the changes to enable gollvm installation,
the package path code in the compiler got thrown off, such that
user-provided "-L" arguments were being mishandled. Specifically, for
"-L /foo" the compiler would look for package export data in /foo/go/7
and /foo/go/7/x86_64-unknown-linux-gnu, but not /foo itself. Fix
things so that we look for the versioned subdirs in the install
directory, but not in user-supplied -L args.

Change-Id: I9af37f504698211756876416e6baa278bb34eec4