gollvm: rework circular function type handling
CL 123738 changes the frontend to explicitly resolve circular
types explicitly, so we don't depend on the order of calls of
placeholder_pointer_type, circular_pointer_type,
set_placeholder_pointer_type, and the construction of the actual
concrete type to resolve circular types.

For circular types, we intercept the cycle with the marker type,
so we won't generate a self-referential type, and link the
marker type (*CPT) to the concrete type.

Adjust test to reflect the new way of API calls by the frontend.

TODO: do the same for circular pointer (non-function) type?

Change-Id: I4264c8913861c7aabae704788b1682cee88bad6b