gollvm: fix buglet in setPlaceholderPointerType
In setPlaceholderPointerType, the placeholder should be set to
equivalent to to_type. We indeed do this for the LLVM type, but
we incorrectly set placeholder to point to to_type in Btype
representation. This CL fixes it.

Move an assertion down after the duplicate check, since once a
placeholder is resolved, it may be equivalent to an anonType in
the sense of Btype::equal. So we only check it for unresolved
placeholders.

Change-Id: I361dc8a46be2b02bc7e94f668406186c98a80d2b