gollvm: rework circular pointer type handing
Following CL 123796, use the concrete type passed by the frontend
for circular pointer types, instead of inferring it ourselves
based on the ordering of a few function calls. Now that we
record the actual concrete element type in the circular pointer
marker type's toType, some routines can be simplified.

Change circularConversionLoadMap_ and circularConversionAddrMap_
to keyed on LLVM types, since LLVM types are unique.

TODO: currently there is no test for the quality of the debug
info. Most of the existing tests are only check that we can
generate debug info (without crash) and they are legal. I
manually checked a few programs with circular types, and the
debug info looks ok to me. The circular type has its concrete
representation in debug info, with the second appearance
represented as *void. For example, "type T *[1]*T" is represented
as *[1]*(*void).

Change-Id: I360aeced5ebfd724dec0818c5e1c0ca77ab778bb