gollvm: fix buglet in -gz
Change the implementation of the "-gz" to be consistent with Clang,
which treats it as a purely assembler-only option. Prior to this patch
the driver was passing "-compress-debug-sections" to both the assembler
and the linker, which is problematic (since the assembler will accept
that option without any args, while the linker always requires an
arg).

Change-Id: I3447a5d94f7a813c9bec61197ccbf95763ade82c
Reviewed-on: https://go-review.googlesource.com/127555
Reviewed-by: Cherry Zhang <cherryyz@google.com>