gollvm: fix compilation flags for runtime/pprof test
Tweak the build rules for the runtime/pprof/check test to compile with
"-static-libgo" and "-fno-inline", to avoid spurious failures in the
TestCPUProfileInlining testpoint (a test that collects a CPU profile and
then tries to verify the correct handling of call stacks within
inlined routines).

With the main Go compiler, expansion of inlined call stacks is done
during the collection run itself, whereas with gccgo the raw PC stack
is captured, with the assumption being that the profile consumer will
do the expansion (by examining the DWARF for each PC sample).

The method used in the test to decode the profile doesn't inspect the
DWARF in the test case binary, meaning that it won't be able to
recover inline info correctly, hence the need for the workaround.

Change-Id: I929cc8b4ff3433b7d9cd628ab5d789502433e135