bridge: unshare deferreturn paths
When a function has defer, the return statements are turned into
calls to runtime.deferreturn. We used to build a single shared
path for all the return points. For the GC support, each return
point may have different set of live variables. For this, we
build distinct path for each return, so we can attach different
stack maps to them. The exceptional path is still shared,
because if an exception happens the frame is essentially dead.

Adjust the unit tests to closer to real code generated by the
frontend, as now we are depending on a particular pattern.

Change-Id: I7712cef62fd498189c4426f58acdc125372ef3a1
Reviewed-on: https://go-review.googlesource.com/c/137758
Reviewed-by: Than McIntosh <thanm@google.com>