unittests: sync tests with changes of call instruction dump in LLVM
LLVM r340519 changes call instruction's dump to print addrspace
if it is not the module's default addrspace, or it cannot find
the module.

In our unit tests, there are two cases. Some tests check the
function's dump. In this case we do have a module, and addrspace
is not printed. Some tests checks a single statement or block. In
this case we have not yet attached the instructions to the
function, so the IR dumper cannot find a module, and thus print
addrspace. Add addrspace to the expected IR dump in this case.

While here, change BackEnd to Backend, for consistency.

Change-Id: I24b6c3b16c4af73159a87660ac7706a919525d93