gollvm: better handling of aggregate-type const conversions
Revamp a portion of the bridge that materializes type conversions for
aggregate constants, so as to enable clean conversions between types
that are structurally identical but are different in their use of
placeholders.

This fixes a compilation problem with the encoding/asn1 package test
exposed as part of the libgo Go 1.11 update.

Change-Id: I0ecf9b23f63c37daeb7c873dcc7e4c8491bed311
Reviewed-on: https://go-review.googlesource.com/c/140117
Reviewed-by: Cherry Zhang <cherryyz@google.com>