bridge: disable inlining functions with defers when GC is enabled
When a function with gets inlined, the outer function may see a
control flow that goes from normal block to landing pad, then go
to normal again (when the inlined function returns to the outer
function).

Currently the liveness code cannot handle it. It treats the
frame dead at a landing pad. Disable inlining such functions for
now.

Change-Id: Ieb9ebf68f6478a0d3467a771ca3efb55f92c38a3