Update patch set 2
Patch Set 2:

I rethought about this. As explained in CL 153439, we should not treat the frame dead at a landing pad. With CL 153439, things seem to work ok without this. I'll abandon this for now. We can revive it if it comes to problem.

Patch-set: 2
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>