gollvm: update FAQ with material on required C/C++ version
Users who build gollvm with older versions of Clang (prior to
5.0) may run into undefined "__get_cpuid_count" symbols at
runtime, due to a not-completely-functional copy of <cpuid.h>
provided by these older compilers.

To help with this, add a section to the instructions on building Gollvm
that calls out the required C/C++ compiler vintage, and add an
item to the FAQ that describes the symptom if an older Clang
is used for the build.

Change-Id: I59f17cc69161e1baf03e7c74926f534d76480290
Reviewed-on: https://go-review.googlesource.com/c/154347
Reviewed-by: Ian Lance Taylor <iant@golang.org>