passes: handle degenerate Phis in stack maps
A degenerate Phi of null pointer is recorded in the stack map as
constant 0. Previously we just ignored constant 0, since it
cannot be a valid pointer bitmap. However, with a type that is
larger than 32*8 bytes, constant 0 may still appear as part of
the multiword bitmap, so it is problematic to just ignore it.
Instead, filter degenerate Phis and don't record them in the
live set.

Change-Id: I2e39393c8eea8d5d1c515445b63d969796591524