passes: improve degenerate Phi detection
It is possible that a group of mutually dependent Phis have a
constant value, like
	a = phi(null, b)
and in a different block
	b = phi(a, null)

The simple algorithm could not detect it.

The codegen preparation pass may reduce this to the constant,
causing problem if it is recorded in the stack map.

Improve degenerate Phi detection to handle this case.

Change-Id: Ie81a78796cfd0f9920411af18758343d886a55e3