passes: mark alloca live when its address is used
When a stack slot's address is taken, it could potentially be
live any point after it. It could also be initialized
"indirectly", like

  tmp = phi(&a, &b)
  *tmp = ...

In this case we didn't see the initialization of either a or b.
But the slot needs to be live. This CL fixes this.

The slot could also be initialized some time later (after a few
statepoints), or never. We don't know for sure. So we need to
pre-zero it, if we don't otherwise know it is initialized.

Also some minor cleanup:
- no need to add pointerless allocas into AddrTakenAllocas.
- delete an unused function (only relevant for relocations).

Change-Id: I44fdb4fb08883bffd51e125aabd6421cbec820e8
Reviewed-on: https://go-review.googlesource.com/c/155761
Reviewed-by: Than McIntosh <thanm@google.com>