bridge: mark write barrier functions GC leaf
Set "gc-leaf-function" attribute for the write barrier
functions. The backend understands this attribute and does not
treats the call sites as statepoints.
(llvm/lib/Transforms/Utils/Local.cpp, callsGCLeafFunction)
This makes the write barriers cheaper (avoid spillings), and
also makes it easier to remove write barriers for stack writes
(not in this CL).

Change-Id: Id5b97275b99c8717d3b9c095231ee0f6e09fe00c