passes: zero ambiguously lived slots at kill sites
Consider the following control flow:

for ... {
	statepoint1 // x is not live here
	{ // inner scope, probably due to inlining
		lifetime_start(x)
		if ... {
			x = ...
			use(&x)
		}
		statepoint2 // x is live here
		lifetime_end(x)
	}
}

Here, x is defined in the inner if block, and has its address
taken. It also has bounded lifetime (e.g. due to inlining). x is
live at statepoint2, as its address may still be used, then it
is killed. At next iteration x is not live at statepoint1, so
its referenced object could be collected there. Then it goes to
the inner scope, and the if branch is not taken, at statepoint2,
x is still holding the old value, which comes back to live
again, which is bad.

To prevent this, we need to zero x at the kill site, so the GC
won't see the old value of x once it is killed.

Change-Id: Iaf6a6d0002e24aa3576d52adb162aae47973d27f
Reviewed-on: https://go-review.googlesource.com/c/156557
Reviewed-by: Than McIntosh <thanm@google.com>