passes: remove write barriers for stack writes
Stack writes don't need write barriers. Also it is bad if a write
barrier is applied to a write on an uninitialized slot.

The frontend already tries not to insert write barriers for stack
writes, but as it is based on AST, it may not identify all the
stack writes. Here, with the SSA-based IR, we can identify the
remaining stack writes and elide the write barriers.

This is not ideal: it doesn't remove the conditional branch that
tests if the write barrier is enabled.

Change-Id: I8d4b299bc2ab9af49e4b4302ca1252edad5a685d