bridge: improve code generation in non-integral pointer mode
Previously, when using non-integral pointers, for local
variables we create allocas in address space 0 and immediately
cast to address space 1. The addrspacecast instructions we
created does not play well with the optimizer (it works
correctly, but much conservative). Also, this makes the debug
info really bad (essentially all the locals are lost).

Observe that loads and stores don't really care the address
space of the pointer. The only time where the address space
matters is when we take the address. Instead of creating
addrspacecast instructions at the time we create the alloca,
emit it only when taking the address.

This way, most of the non-address-taken allocas will not have
addrspacecast, so it can be SSA'd. The address-taken ones cannot
be SSA'd anyway, so the addrspacecast doesn't hurt much.

As the addrspacecast instruction generation is delayed, it makes
the deferreturn sequence slightly more complicated, as there is
now dependency between the instructions in the sequence. Handle
this in the instruction cloning code in genDeferReturn.

Adjust the call sequence generation as the temporaries for
passing the arguments and results are now in address space 0.

Change-Id: I8400fc4c0b19b6038b70a6949eb2447bc3202582
Reviewed-on: https://go-review.googlesource.com/c/156677
Reviewed-by: Than McIntosh <thanm@google.com>