passes: delete lifetime markers on Phis involving ambiguously live slots
If a Phi (presumably holding an alloca) has lifetime markers and
it gets assigned the same stack slot as an alloca, it could
shorten the lifetime of the slot unwantedly, causing the slot
being reused while still recorded as live. Prevent this from
happening by removing the lifetime markers on Phis containing
ambigously live allocas.

Change-Id: Ia326ba82b6225f18ea3cc7b8652066e53abe7cc4