bridge: fix bug in creating global variable
In the path that we first see an external declaration then a
definition, the definition should override the declaration, and
we should set it to not external. It seems that
setExternallyInitialized(false) is not enough. We need to set
the initializer.

However, this is a bug in setting the initializer. We have two
variables, initializer and init. I think they suppose to be one.
This CL unifies them. With this, the definition actually
overrides the declaration.

Change-Id: I2130c29970c5c440895bb15bee303186f0ea5490