passes: fix a memory error in statepoint pass
For "bad loads", we delete them when they are replaced. However,
it turns out that they may be still referenced by the liveness
maps. We do filter out them when we compute the live sets of
basic blocks. But we didn't do this for computing the live set of
an instruction. This reference can cause memory errors in the
compiler.

This CL changes the handling of bad loads to a more systematic
way. We perform filtering of the bad loads in a single place for
all live sets, in the same place as filtering constants. Also,
we move the deletion later, after the live sets are installed to
the IR. This avoids the memory error even if the bad loads are
still somehow referenced. If they are, the deletion will assert.

While here, unify the handling of bad loads and bad Phis.

While here, realize that Phi->hasConstantValue does not
necessarily return whether the Phi is constant. It just returns
if the Phi is degenerate. As we want to check for constants, add
an extra check.

Change-Id: Ib605e75c71646653e1472bde7d7f2c68b6e481bf