passes: handle alloca kill and def in the same block
It is possible that an alloca is killed then defined in the same
block. One example is loop unrolling. For example, a loop

b1:
  def x
  goto b2
b2:
  kill x
  goto b1

after unrolling:

b1.1:
  def x
  goto b2
b2.1:
  kill x
  // continue on b1.2
  def x
  goto b2.2
b2.2
  kill x

b2.1 contains a kill and a def. x is defined at the end of b2.1.
We should record this information in the alloca def set.

Currently, we subtract the kill set from the def set. This
information is lost. Fix this by removing x from the kill set
when we see a def.

Change-Id: If8eef510f0fa14c0d7f5bb0dcd840dac19fedd4c
Reviewed-on: https://go-review.googlesource.com/c/163202
Reviewed-by: Than McIntosh <thanm@google.com>