gollvm: add a Go specific "sret" attribute
With -O3, the argument promotion pass can optimize a StructRet
argument to nonnull noalias. The StructRet attribute is lost.
However, the statepoint pass uses this attribute to detect if a
stack slot is initialized at a call.

Introduce a Go specific "sret" attribute and let the statepoint
pass check for it. This attribute is preserved during the
optimization.

Now precise stack scan works with -O3.

Change-Id: I3cbd8bd8ca7421fd2d4900e73d50edd7c963d56c