bridge: make an extra copy when passing an alloca as a byval arg
When passing an alloca as a byval arg, the LLVM optimizer seems
to think it uses the address, which prevents it from being SROA'd
or registerized. Work around this by generating an extra copy,
which is only used in this call, and the original alloca could be
SROA'd if otherwise it can be. Currently we only do this when it
is an alloca (or an offset from it), and the type is small.

Change-Id: If43f9fe6691c64202d7affd483c277a33fa8c926