gollvm: fix bug in call->invoke instruction rewriting
During the control flow generation pass in gollvm, there is a pass
that rewrites call instructions to invoke instructions in situations
where this is needed due to the presence of a defer. The rewriter
wasn't properly handling the case where a given Bexpression was
inheriting the LLVM value of a child expression, and as a result the
compiler was crashing when trying to access a stale (deleted) LLVM
value.

Fixes golang/go#31388.

Change-Id: I6abc5ec87b3f63d2974126355a9d8fe18ec5ed1f