gollvm: tweak cmake rules for aeshash.c to include correct flags
The libgo runtime source aeshash.c contains AES/SSE3-specific
intrinsics that require the arch features not ordinarily present for
default x86_64 compilation. This source file contains a "#pragma GCC
target" directive that asks for the features, but that pragma is not
supported by clang. Prior versions of clang would go ahead and inline
in the AES intrinsics anyhow, but never ones will issue an error.
To work around this, change the compile flags for aeshash.c to include
"-maes -mssse3".

Change-Id: I638059cd429c5e1157deccd52d76cceea5d3c087