bridge: handle conversion expression applied on AuxT
We use AuxT to wrap an LLVM type used in LLVM intrinsics. A value
of such type could be fed into a conversion expression, like
uint64(__builtin_ctz(x)). materializeConversion needs to handle
this case.

For AuxT, we don't know if it is signed or not. For now we simply
assume it is signed. In most cases it doesn't matter. If this
turns to be an issue, we should define the correct Btype for
intrinsics.

Change-Id: If60c5ccb3e6aa1c4bf0cb985abbbce9f019456e5