bridge: make nil pointer dereference "defined"
Generally the LLVM backend treats nil pointer dereference as
undefined behavior. In Go, this is actually defined as triggering
a run-time panic. Set "null-pointer-is-valid" attribute to tell
the backend not to rewrite nil pointer dereference.

Due to the lack of non-call exceptions, this is not enough for us
to remove the explicit nil checks. But this is enough to make
sure intrinsified atomic operations, when the input is nil, will
seg fault.

Change-Id: Icd26acfcf94247bbb0213278842f8fcc3701bbc2