bridge: support implicit_variable_reference
implicit_variable_reference is just a reference of an external
global variable.

Relax some assertions in creating global variable. When we see a
definition and a declaration of the same global variable in the
same package, the definition wins. Here we don't require some
attributes (e.g. isConstant) match, just overwrite them. This is
because some attributes may be not known when we create a
declaration, e.g. implicit_variable_reference. And these
attributes don't really matter for an external declaration.

Change-Id: I981061880326dad10e7b011074c330a36c461ecb