gollvm: adjust symbol linkage recipe for 'only_inline' functions
Change the recipe used to compute symbol linkage for functions that
the front end has marked "only_inline".  Up to this point all inline
functions emitted by the front end were marked as non-visible; now
(at the moment) all functions are marked as visible. The new recipe
is roughly:

  (!visible ? InternalLinkage :
    (only_inline ? AvailableExternallyLinkage :
                   ExternalLinkage)

Includes an update to the unit test for this attribute as well.

Change-Id: I5ebc5051ded4a9ce1bf96dcbf8ca6d074f8769ca
Reviewed-on: https://go-review.googlesource.com/c/gollvm/+/180478
Reviewed-by: Cherry Zhang <cherryyz@google.com>