driver: compile runtime package with nil check removed
Panics in the runtime are uncatchable throws, which will crash
the program anyway. We don't have to worry about if it panics at
the exact memory state as the program order.

With the previous CL, the effect of this CL is allowing the
backend to reorder instructions across a potentially faulting
instruction. The faulting instruction is not removed. So if there
is actually a bug in the runtime causing nil check to fail, it
still crash the program.

Change-Id: Ic87b0fb351be234b3b6727ecb97093257c07a7cf