bridge: use LLVM value type for building GEP
Currently when building a field expression we pass the
Bexpression's type to the IR builder. However, in some cases this
type is not exactly the same type as the LLVM value type -- one
is a named struct type, one is not. This causes the IR builder to
fail. Work around this by always using the LLVM value's type.

Change-Id: I5ad91c78180693305cd28cc397e3625e2ab0c7c7