gollvm: additional debugging routines
DO NOT SUBMIT

Add in a couple of new helper functions for debugging. This includes
dump routines for labels, and a new routine for diagnosing why a given
placeholder type was not fully converted.

NB: not sure if these are worth submitting; I've used them a couple of
times for tracking down bugs, but it's unclear whether they are of
general interest.

Change-Id: I4d8c72c2e5903080ef6caf25c82c6a8c93aca36d