gollvm: fix bug in materialize of conditional expression
Fix a bug in the code that materializes conditional expressions
(triggered by the recent gofrontend changes to bounds checking). Avoid
creating a temporary variable for the conditional if either the Btype
of the expression itself is void, or the Btype of the "else"
expression is void, otherwise we may wind up trying to create an
alloca with void type.

Fixes golang/go#33964.

Change-Id: I9a2ea902026d0bd8d56ead9dfcc22685ef6e7a62