gollvm: refactor cabiOracle and Llvm_backend class
Currently, some code of gollvm is written specifically for amd64, such as classes
cabiOracle and Llvm_backend. To make it easier for gollvm to extend to multiple
architectures, this patch mainly makes the following modifications:
1, Adds a llvm::CallingConv::ID type parameter to the constructor of class Llvm_backend,
   so that we can generate different backend according to this parameter.
2, Refactors class cabiOracle with class inheritance. Firstly we add a pure virtual
   class CABIOracleArgumentAnalyzer, which contains the common architecture independent
   interfaces. Then implemente a sub-class CABIOracleX86_64_SysV for amd64, it inherits
   class CABIOracleArgumentAnalyzer. Finally add a pointer data member of class
   CABIOracleArgumentAnalyzer into class CABIOracle, and rewrite part of the public
   functions of class CABIOracle with this pointer. The pointer will be assigned to
   different sub-class instance object according to the calling convention value of type
   manager, so we can have different ABI implementation with class cabiOracle.
With this change, we can easily port gollvm to another architecture by implementing
a sub-class of class CABIOracleArgumentAnalyzer.

This CL doesn't modify the unit test cases, we'll do this work in the follow up patches.

Change-Id: I40ab0cfc0012f404f3d237844d4731fc5623b686