Update patch set 1
Patch Set 1:

Thanks for sending this! I will begin looking it over.

For modifications of this size, it would help the reviewers a lot if the changes can be split into multiple patches; if you'd be willing to do this it would make it a lot easier.

For example, the switch from moving away from hard-coding the calling convention to passing it in as a parameter could be done as a separate initial patch (upon which other patches can depend).

Patch-set: 1
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>