Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks for sending this! I will begin looking it over.
> 
> For modifications of this size, it would help the reviewers a lot if the changes can be split into multiple patches; if you'd be willing to do this it would make it a lot easier.
> 
> For example, the switch from moving away from hard-coding the calling convention to passing it in as a parameter could be done as a separate initial patch (upon which other patches can depend).

Thanks Than! At the beginning, I planned to split it into multiple patches, but I hope you see the whole change, and the other is that I feel that rebase is a bit tedious, so I didn't split it. But it does cause a lot of trouble for the reviewer, I will split it into multiple patches. Maybe you can wait for me to split it, then review, thank you.

Patch-set: 1