gollvm: add check for '-fsplit-stack' option
On some platforms such as linux arm64, gcc and ld.gold don't support
'-fsplit-stack' option at present, but currently gollvm doesn't check this,
and enters the normal compilation process, then fails when linking. This CL
adds a check for this option when parsing command line options, so that if
'-fsplit-stack' is not supported, the program can exit earlier and report
the corresponding error message.

Change-Id: I41f74435f790375c4dedd0c5e544124e67d62087
Reviewed-on: https://go-review.googlesource.com/c/gollvm/+/210323
Reviewed-by: Than McIntosh <thanm@google.com>