Update patch set 1
Patch Set 1:

Ah, ok, I think I understand now. 

In libgo the __wrap_pthread_create function is defined in the split-stack support code (e.g. morestack).

So it sounds as though what's needed is to make the injection of "--wrap" dependent on whether we're using split-stack or not (as opposed to having it on or off unconditionally).

Patch-set: 1
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>