Update patch set 1
Patch Set 1:

(1 comment)

> Looks good to me overall. Should this depend on the actual -f[no-]split-stack flag passed at invocation?

Well, that's an interesting question. I guess my feeling is that it would be better not to look at the flag, since it is hard to tell on the linker invocation whether everything feeding into the link was compiled consistently (could easily be a mix of with/without).

Patch-set: 1