Update patch set 1
Patch Set 1:

(5 comments)

This looks good overall. 

Curious if you were able to do testing with https://github.com/thanm/cabi-testgen, and/or if it was useful (or if it needs enhancements/fixes).

Thanks for contributing this, I am sure it was a good deal of work to write.

Patch-set: 1
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>